Use subdirectories for ts code
