+++ src/apiImplementation.ts
@@ -1,4 +1,4 @@
-import { RHelp } from './rHelp';
+import { RHelp } from './helpViewer';
 import { RExtension } from './api';
 export class RExtensionImplementation implements RExtension  {
     public helpPanel?: RHelp;
