+++ src/helpViewer/helpProvider.ts
@@ -7,7 +7,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as os from 'os';
 
-import * as rHelp from './rHelp';
+import * as rHelp from '.';
 
 export interface RHelpProviderOptions {
 	// path of the R executable
